Added more flow-control tests using k3po nukleus transport.
